+++ src/pulumi/core-dtap/analytics/datafactory/__init__.py
@@ -3,6 +3,7 @@ from project_config import platform_outputs
 
 platform_outputs["analytics"]["datafactory"] = {"factories": {}}
 
+from . import datafactories
 from . import integrated_integration_runtime
 from . import orchestration
 from . import orchestration_linked_services